+++ backend/src/main/java/reviewme/review/service/ReviewListLookupService.java
@@ -13,6 +13,7 @@ import reviewme.review.service.mapper.ReviewListMapper;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.domain.exception.ReviewGroupNotFoundException;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
+import reviewme.util.PageSize;
 
 @Service
 @RequiredArgsConstructor
